.cursor/ativo_real/backend/requirements.txt
Django>=5.0
Djangorestframework>=3.15
psycopg2-binary>=2.9
django-filter>=23.5
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
djangorestframework-gis>=1.1
